requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
psycopg2==2.9.6
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
six==1.16.0
spotipy==2.22.1
urllib3==1.26.15
